refactor(backend): move the concurrency test helpers into helpers/
Review flagged handleLater being defined twice. createBarrier and
createValueBarrier were duplicated across the same two files for the same
reason, so all three move to __tests__/helpers/concurrency.ts alongside the
existing fixtures. Net -68 lines.

deferred-rejection.test.ts now imports the shipped handleLater instead of
re-declaring a copy, so it pins the real implementation.
