+++ packages/backend/src/__tests__/board-merge-tombstone.test.ts
@@ -9,6 +9,7 @@ import { boardPresenceMutations } from '../graphql/resolvers/board-presence/muta
 import { findChosenBoardForSerial } from '../graphql/resolvers/board-presence/shared';
 import { lockBoardSerialWrite } from '../graphql/resolvers/board-serial-write-lock';
 import { seedAuroraCatalogFixtures } from './helpers/board-catalog-fixture';
+import { createBarrier, createValueBarrier, handleLater } from './helpers/concurrency';
 
 /**
  * Real-DB coverage for issue #3407's merge-tombstone following and the
@@ -155,40 +156,6 @@ async function serialPointerUuid(userId: string, serial: string): Promise<string
     `),
   );
   return rows[0]?.board_uuid ?? null;
-}
-
-function createBarrier(): { promise: Promise<void>; release: () => void } {
-  let release = (): void => undefined;
-  const promise = new Promise<void>((resolve) => {
-    release = resolve;
-  });
-  return { promise, release };
-}
-
-function createValueBarrier<Result>(): { promise: Promise<Result>; release: (result: Result) => void } {
-  let release = (_result: Result): void => undefined;
-  const promise = new Promise<Result>((resolve) => {
-    release = resolve;
-  });
-  return { promise, release };
-}
-
-/**
- * Attach an inert rejection handler to a promise this test asserts on later.
- *
- * These tests deliberately leave a promise in flight across several `await`s.
- * If it rejects during one of them, Node reaches its unhandled-rejection
- * checkpoint before the assertion further down can attach a handler — vitest
- * counts that under `Errors` and exits 1 with every test still passing
- * (issue #4488, main run 31880202330).
- *
- * The promise itself is untouched: still rejected, still assertable. The
- * `await expect(...).rejects.toThrow(...)` or bare `await` below still runs,
- * still sees the same rejection, and still fails on the wrong error — so this
- * cannot hide a defect. Its contract is pinned in deferred-rejection.test.ts.
- */
-function handleLater(promise: Promise<unknown>): void {
-  void promise.catch((): void => undefined);
 }
 
 async function waitForAdvisoryWaitBlockedBy(blockingPid: number): Promise<number> {
